feat: StationName 추가
 Station.name의 validate 책임 분할
