+++ src/main/java/wooteco/subway/admin/domain/Station.java
@@ -1,39 +1,32 @@
 package wooteco.subway.admin.domain;
 
-import java.util.Objects;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Embedded;
+import wooteco.subway.admin.domain.vo.StationName;
 
 public class Station {
 
     @Id
     private Long id;
-    private String name;
+    @Embedded.Nullable
+    private StationName stationName;
 
-    public Station(String name) {
-        validateName(name);
-        this.name = name;
+    public Station() {
     }
 
-    private void validateName(String name) {
-        if (Objects.isNull(name)) {
-            throw new IllegalArgumentException("Station.name은 null이 올 수 없습니다.");
-        }
-        if (name.isEmpty()) {
-            throw new IllegalArgumentException("Station.name은 빈 값이 올 수 없습니다.");
-        }
-        if (name.matches(".*[ ].*")) {
-            throw new IllegalArgumentException("Station.name은 공백이 포함 될 수 없습니다.");
-        }
-        if (name.matches(".*[0-9].*")) {
-            throw new IllegalArgumentException("Station.name은 숫자가 포함 될 수 없습니다.");
-        }
+    public Station(StationName stationName) {
+        this.stationName = stationName;
+    }
+
+    public Station(String stationName) {
+        this(new StationName(stationName));
     }
 
     public Long getId() {
         return id;
     }
 
-    public String getName() {
-        return name;
+    public StationName getStationName() {
+        return stationName;
     }
 }
